requirements.txt
flask==3.1.1
flask-cors==6.0.1
numpy==1.26.4
matplotlib==3.7.5
tensorflow==2.18.0
Pillow==11.2.1
tf-keras-vis==0.8.7
gunicorn==21.2.0 